+++ test/navigation/NavigatorSpec.scala
@@ -17,7 +17,7 @@
 package navigation
 
 import base.SpecBase
-import controllers.routes
+import controllers.monthlyreturns
 import pages._
 import pages.monthlyreturns.*
 import models._
@@ -57,7 +57,11 @@ class NavigatorSpec extends SpecBase {
       "must go from a page that doesn't exist in the route map to CheckYourAnswers" in {
 
         case object UnknownPage extends Page
-        navigator.nextPage(UnknownPage, NormalMode, UserAnswers("id")) mustBe routes.CheckYourAnswersController
+        navigator.nextPage(
+          UnknownPage,
+          NormalMode,
+          UserAnswers("id")
+        ) mustBe monthlyreturns.routes.CheckYourAnswersController
           .onPageLoad()
       }
     }
@@ -69,7 +73,7 @@ class NavigatorSpec extends SpecBase {
           DateConfirmNilPaymentsPage,
           CheckMode,
           UserAnswers("id")
-        ) mustBe routes.CheckYourAnswersController.onPageLoad()
+        ) mustBe monthlyreturns.routes.CheckYourAnswersController.onPageLoad()
       }
 
       "must go from InactivityRequestPage to CheckYourAnswers Page in CheckMode" in {
@@ -77,7 +81,7 @@ class NavigatorSpec extends SpecBase {
           InactivityRequestPage,
           CheckMode,
           UserAnswers("id")
-        ) mustBe routes.CheckYourAnswersController.onPageLoad()
+        ) mustBe monthlyreturns.routes.CheckYourAnswersController.onPageLoad()
       }
 
       "must go from ConfirmEmailAddressPage to CheckYourAnswers Page in CheckMode" in {
@@ -85,13 +89,17 @@ class NavigatorSpec extends SpecBase {
           ConfirmEmailAddressPage,
           CheckMode,
           UserAnswers("id")
-        ) mustBe routes.CheckYourAnswersController.onPageLoad()
+        ) mustBe monthlyreturns.routes.CheckYourAnswersController.onPageLoad()
       }
 
       "must go from a page that doesn't exist in the edit route map to CheckYourAnswers" in {
 
         case object UnknownPage extends Page
-        navigator.nextPage(UnknownPage, CheckMode, UserAnswers("id")) mustBe routes.CheckYourAnswersController
+        navigator.nextPage(
+          UnknownPage,
+          CheckMode,
+          UserAnswers("id")
+        ) mustBe monthlyreturns.routes.CheckYourAnswersController
           .onPageLoad()
       }
     }
